Added aux scaler to the analyzer. It's done in a fully backwards compatible way. When analazing new files, which have the aux scaler, the data are unpacked into a new tree 't8'. When analyzing old files w/o the aux scaler present, it's simply ignored, though a tree called 't8' will be present in the ROOT files created by mid2root, they will have no events in them.
Also added a new feature to midas::Database, the CheckPath() function, which allows to
silently check if a path is present in the DB. This was motivated by the scaler changes,
so I could easily figure out if a midas file was pre- or post-scaler addition by looking
for the "/Equipment/AuxScaler/Settings/Rate" ODB array, which will only be there _after_
the aux scaler was added.

Note that the aux scaler is a new IO32 module which serves only to give us 16 additional
scaler channels, the primary purpose being to use them for the photodiodes.